Move the diff package to the shared location
