+++ pkg/app/piped/cloudprovider/kubernetes/BUILD.bazel
@@ -19,10 +19,10 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/app/piped/chartrepo:go_default_library",
-        "//pkg/app/piped/diff:go_default_library",
         "//pkg/app/piped/toolregistry:go_default_library",
         "//pkg/cache:go_default_library",
         "//pkg/config:go_default_library",
+        "//pkg/diff:go_default_library",
         "//pkg/git:go_default_library",
         "//pkg/model:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
