+++ pkg/app/piped/planner/kubernetes/BUILD.bazel
@@ -12,9 +12,9 @@ go_library(
         "//pkg/app/piped/cloudprovider/kubernetes:go_default_library",
         "//pkg/app/piped/cloudprovider/kubernetes/resource:go_default_library",
         "//pkg/app/piped/deploysource:go_default_library",
-        "//pkg/app/piped/diff:go_default_library",
         "//pkg/app/piped/planner:go_default_library",
         "//pkg/config:go_default_library",
+        "//pkg/diff:go_default_library",
         "//pkg/model:go_default_library",
         "@org_uber_go_zap//:go_default_library",
     ],
