+++ pkg/diff/diff_test.go
