+++ pkg/diff/renderer.go
