+++ pkg/diff/renderer_test.go
