+++ pkg/diff/result.go
