+++ pkg/diff/result_test.go
