+++ pkg/diff/testdata/complex-node.yaml
