+++ pkg/diff/testdata/has_diff.yaml
