+++ pkg/diff/testdata/ignore_adding_map_keys.yaml
