+++ pkg/diff/testdata/no_diff.yaml
